Connection to quickbooks online is ready. Tomorrow I will write a script to inactivate all customers that don't have an invoice or sales receipt in the last 100 days. This is something I have been trying to do manually in the last few weeks and have made little progress. Managing customers is extremely difficult when you have short term property rentals where people have balances, but most customers are no longer active.
Then once I have a concise list of likely active customers, I can sit down with the manager and see who is on my list that isn't on their paper list. Then hopefully get away from the paper list which requires secret knowledge, because names are not standardized e.g. company name only or last name only and hard to find in addition to having a lack of history and not allowing past due balances to be described.

Then once this is done, I'd like to look into grouping customers by location in the flea market, creating invoices automatically and hopefully figuring out what each person has purchased besides an amount, because amounts change and people get credits and then it makes it difficult for someone new to get the full normal amount after credits end or even just knowing what is supposed to be known.

There's a lot of the rental process that needs to be formally defined, and made so much simpler by going through the trouble of formally defining which then allows processes to be automated.